multiple changes.
Fixed the height of base of calculate to 100% and fixed it’s bottom
padding.
Fixed some of the divs in the index file.

fixed the screed display of the calculator.  as the larger numbers and
calculation chains were expanding the screens. There is now a limit to
the size of the number you can build as I changed some of the if
functions permissions for number pushing.

I removed the SCSS files as I have decided not to use them.

I ran the build function on the project to recompile it.